Rename `crds` to `printcrds`
