+++ cmd/experimental/kjobctl/pkg/cmd/printcrds/printcrds_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package crds
+package printcrds
 
 import (
 	"bytes"
